requirement.txt
langchain==0.1.10
langchain_community==0.0.25
langchain_core==0.1.28
langchain_openai==0.0.8
python-dotenv==1.0.1
streamlit==1.27.2
streamlit==1.31.1
tqdm==4.66.1
tqdm==4.66.2